# doc/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Library API",
    "description": "With this API you can create and safe your books",
    "contact": {
      "email": "vmartins@daitan.com"
    }
  },
  "host": "localhost:3333",
  "basePath": "/livros",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/paginate/{pages}": {
      "get": {
        "tags": [
          "Library"
        ],
        "summary": "Get a limited number of books",
        "description": "Get a limited number of books",
        "parameters": [
          {
            "name": "pages",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Number of elements"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
...
}